MIDI-116: Dashboard

+++ models/modules/masking.py
@@ -16,7 +16,7 @@ def random_masking(x: torch.Tensor, masking_ratio: float):
     x_masked = torch.gather(x, dim=1, index=ids_keep.unsqueeze(-1).repeat(1, 1, D))
 
     # generate the binary mask: 0 is keep, 1 is remove
-    mask = torch.ones([N, L], device=x.device)
+    mask = torch.ones([N, L], device=x.device, dtype=int)
     mask[:, :len_keep] = 0
     # unshuffle to get the binary mask
     mask = torch.gather(mask, dim=1, index=ids_restore)
